test(datalab): ♻️ move tests for issue managers into their respective test modules
